requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-SQLAlchemy
Jinja2
Werkzeug
requests
gunicorn
psycopg2-binary
pyopenssl